feature: Created Project context.
It is containing all of the functions of the presented project,
and oragnized the code accordingly to changes so it won't effect
the functuality.
In addition, I have also added animation the container settings button,
and added close context menu in item when also clicking outside of it.
also moved all of the project types into one file.